chore: move components to files
